rename hooks -> plugin
